re_types: clean up
